# specs/002-fullstack-web-crm/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TeamFlow API
  description: |
    RESTful API for TeamFlow Phase 2 - Agency CRM with task management,
    time tracking, and profitability reporting.

    **Authentication**: JWT Bearer tokens (httpOnly cookies)
    **Multi-Tenant**: All requests scoped to agency_id from JWT claims
    **Base URL**: https://api.teamflow.example.com/v1

  version: 1.0.0
  contact:
    name: TeamFlow API Support
    email: api@teamflow.example.com
servers:
  - url: http://localhost:8000/v1
    description: Local development
  - url: https://api.teamflow.example.com/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register new agency and user
      description: |
        Creates a new agency and first user (admin). Returns JWT token.
        This is the only endpoint that doesn't require authentication.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agency_name, email, password, name]
              properties:
                agency_name:
                  type: string
                  minLength: 2
                  maxLength: 255
                  example: "Creative Studio Inc"
                email:
                  type: string
                  format: email
                  example: "admin@creativestudio.com"
                password:
                  type: string
                  minLength: 8
                  example: "SecurePass123!"
                name:
                  type: string
                  minLength: 2
                  maxLength: 255
                  example: "Jane Admin"
      responses:
        '201':
          description: Registration successful
          headers:
            Set-Cookie:
              description: JWT token in httpOnly cookie
              schema:
                type: string
                example: "auth_token=eyJ...; Path=/; HttpOnly; Secure; SameSite=Strict"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [auth]
      summary: Login with email and password
      description: |
        Authenticates user and returns JWT token.
        Token includes agency_id for tenant scoping.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@agency.com"
                password:
                  type: string
                  example: "MyPassword123!"
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users:
    get:
      tags: [users]
      summary: List team members
      description: |
        Returns all active users in the same agency.
        Excludes soft-deleted users (active=false).
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  total:
                    type: integer
                    example: 5

    post:
      tags: [users]
      summary: Create team member
      description: |
        Creates a new team member in the same agency.
        Auto-generates a secure temporary password if not provided.
        Password expires in 7 days by default.
        Requires current_user to be admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserReadWithTempPassword'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Only admins can create users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists in agency
  /users/{user_id}:
    parameters:
      - $ref: '#/components/parameters/UserId'

    get:
      tags: [users]
      summary: Get user by ID
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [users]
      summary: Update user
      description: |
        Partial update of user fields.
        Only admins can modify role and is_project_manager.
        Users can update their own name and email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Insufficient permissions
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [users]
      summary: Delete user (soft delete)
      description: |
        Soft deletes user by setting active=false.
        User's assigned tasks become unassigned.
        Cannot delete the last admin in the agency.
        Requires current_user to be admin.
      responses:
        '204':
          description: User deactivated
        '403':
          description: Cannot delete last admin or insufficient permissions
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # TIME ENTRIES ENDPOINTS
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from httpOnly cookie or Authorization header
  parameters:
...
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Validation failed"
        details:
          type: object
    AuthResponse:
      type: object
      required: [access_token, token_type, user]
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [id, email, name, role, agency_id]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          minLength: 1
          maxLength: 100
        role:
          type: string
          enum: [admin, member, client]
          default: member
        agency_id:
          type: string
          format: uuid
        active:
          type: boolean
          default: true
          description: Soft delete flag
        is_project_manager:
          type: boolean
          default: false
          description: Can create and manage projects
        password_expires_at:
          type: string
          format: date-time
          nullable: true
          description: Temporary password expiration
        must_change_password:
          type: boolean
          default: false
          description: Force password change on login
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    UserCreate:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          description: Optional password, auto-generated if not provided
        role:
          type: string
          enum: [admin, member, client]
          default: member
        is_project_manager:
          type: boolean
          default: false
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, member, client]
          description: Admin-only field
        is_project_manager:
          type: boolean
          description: Admin-only field
    UserReadWithTempPassword:
      allOf:
        - type: object
          required: [id, email, name, role, agency_id, temporary_password]
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            name:
              type: string
            role:
              type: string
              enum: [admin, member, client]
            agency_id:
              type: string
              format: uuid
            active:
              type: boolean
            is_project_manager:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
              nullable: true
            temporary_password:
              type: string
              description: Auto-generated temporary password (shown once on creation)
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'